+++ src/components/tabs/tabs.js
@@ -70,6 +70,14 @@ import classNames from 'classnames';
  */
 class Tabs extends React.Component {
 
+  /**
+   * Store the window object as property.
+   *
+   * @property _window
+   * @type {Object}
+   */
+  _window = window
+
   static propTypes = {
 
     /**
@@ -124,7 +132,15 @@ class Tabs extends React.Component {
      * @property titleSize
      * @type {String}
      */
-    titleSize: React.PropTypes.string
+    titleSize: React.PropTypes.string,
+
+    /**
+     * Emitted when a tab header is clicked
+     *
+     * @property onTabClick
+     * @type {Func}
+     */
+    onTabClick: React.PropTypes.func
   }
 
   static defaultProps = {
@@ -182,8 +198,24 @@ class Tabs extends React.Component {
     if (this.props.initialTabId) {
       initialSelectedTabId = this.props.initialTabId;
     } else {
+      let hash = this._window.location.hash.substring(1);
+
       if (Array.isArray(this.props.children)) {
-        initialSelectedTabId = compact(this.props.children)[0].props.tabId;
+        let children = compact(this.props.children),
+            useHash = false;
+
+        if (hash) {
+          for (let index in children) {
+            let child = children[index];
+
+            if (child.props.tabId == hash) {
+              useHash = true;
+              break;
+            }
+          }
+        }
+
+        initialSelectedTabId = useHash ? hash : children[0].props.tabId;
       } else {
         initialSelectedTabId = this.props.children.props.tabId;
       }
@@ -211,7 +243,13 @@ class Tabs extends React.Component {
    * @param {Event} ev Click Event
    */
   handleTabClick = (ev) => {
-    this.setState({ selectedTabId: ev.currentTarget.dataset.tabid });
+    let tabid = ev.target.dataset.tabid;
+    this._window.location = `#${tabid}`;
+    this.setState({ selectedTabId: tabid });
+
+    if (this.props.onTabClick) {
+      this.props.onTabClick(tabid);
+    }
   }
 
   /**